fixing build issues

+++ my-app/actions/summary-actions.ts
@@ -12,12 +12,16 @@ export async function DeleteSummary(summaryId:string){
             redirect('/sign-in');
             throw new Error("user not found");
         }
+        if (!sql) {
+            throw new Error('Failed to get database connection');
+        }
         const res = await sql`DELETE FROM pdf_summaries WHERE id=${summaryId} AND user_id=${userId} RETURNING id`;
         //revalidate path
         if(res.length>0){
-        revalidatePath(`/dashboard`);
-        return{success:true};
-    }
+            revalidatePath(`/dashboard`);
+            return{success:true};
+        }
+        return{success:false};
     }
     catch(error){
